+++ __test__/client.test.jsx
@@ -1,0 +1,3 @@
+// import { shallow, mount, render } from 'enzyme';
+
+// const wrapper = shallow(<Foo />);
