+++ __test__/server.test.js
@@ -1,18 +1,9 @@
+/* eslint-disable no-undef */
 const supertest = require('supertest');
-const app = require('./server/index.js');
-const db = require('./database/index.js');
+
+const app = require('../server/index.js');
 
 const request = supertest(app);
-
-const dbQuery = (sql, cb = () => {}) => {
-  db.con.query(sql, (err, result) => {
-    if (err) {
-      console.log(`Data retrieval error: ${sql}`, err);
-    }
-    cb(result);
-  });
-};
-
 
 // server tests
 describe('/api/steps/1', () => {
@@ -50,29 +41,4 @@ describe('/api/videos/1', () => {
     expect(response.body[0]).toHaveProperty('url');
     expect(Array.isArray(response.body)).toEqual(true);
   });
-});
-
-
-
-// database tests
-
-describe('Database Test', () => {
-  beforeAll (async () => {
-    await db.con.query('USE bonappetit;');
-  });
-
-  test('Name should be Cinnamon-Date Sticky Buns when id = 1', async ()=> {
-    await dbQuery('SELECT name FROM recipes WHERE id = 1;', (result) => {
-      const response = result[0].name;
-      expect(response).toEqual('Cinnamon-Date Sticky Buns');
-    });
-  });
-
-  test('Should contain 100 unique recipes', async ()=> {
-    await dbQuery('SELECT * FROM recipes;', (result) => {
-      const response = result.length;
-      expect(response).toEqual(100);
-    });
-  });
-
 });
